+++ packages/svelte/CHANGELOG.md
@@ -5,6 +5,7 @@
 - add `selectable`, `deletable` and `draggable` to node and edge props
 - add `parentId` to node props
 - add `on:edgemouseenter` and `on:edgemouseleave` event handler
+- fix deselection of edges
 
 ## 0.1.3