+++ packages/svelte/CHANGELOG.md
@@ -7,6 +7,7 @@
 - add `on:edgemouseenter` and `on:edgemouseleave` event handler
 - fix deselection of edges
 - remove pointer events from panel when user selection is active
+- fix viewport initialization with user viewport 
 
 ## 0.1.3